+++ objsonl/builder.py
@@ -3,14 +3,15 @@ from typing import List
 from typing import Optional
 
 import pandas as pd
-from objsonl.model import Batch
-from objsonl.model import Body
-from objsonl.model import Detail
-from objsonl.model import Line
-from objsonl.model import Message
-from objsonl.model import SystemMessage
-from objsonl.model import UserMessage
-from objsonl.model import UserMessageWithImage
+
+from objsonl.model.input import Batch
+from objsonl.model.input import Body
+from objsonl.model.input import Detail
+from objsonl.model.input import Line
+from objsonl.model.input import Message
+from objsonl.model.input import SystemMessage
+from objsonl.model.input import UserMessage
+from objsonl.model.input import UserMessageWithImage
 
 __ALL__ = [
     "JsonlBuilder",
